requirements.txt
Django==1.6.2
-e git+ssh://git@github.com/egonSchiele/Django-EditArea.git@f27131d2a210e4dd32cffa3f29d73fcc2477e1af#egg=Django_EditArea-master
MySQL-python
South==0.8.4
Unidecode==0.04.14
django-extensions==1.3.3
django-reversion==1.8.0
django-taggit==0.11.2
jsonfield==0.9.20
python-slugify==0.0.7
six==1.5.2
wsgiref==0.1.2